+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 1.0.0
 
 - Initial version.
+
+## Unreleased
+- Fix handling of very large sizes in string.packsize parser